Skip setup if in SSH session.
If the notification are enabled, every command that takes longer than
10 seconds (even if in the same shell) will trigger the notify-send
command and will hug the terminal until cancelled (ctrl-C). It happens
also when using a file editor like vi.

It would be possible to use the notification in an SSH session if an X
server is set up correctly. Maybe another check could be added to see if
there is an X server available or if the X redirection is enabled
because totally disabling the notification.

Change-Id: Ib3b7324b8054f7a87f7d2d11c2f945b43485f30d